### 2025-02-28 #7860 @editorialbot
[PRE REVIEW]: Heatwave Diagnostics Package: Efficiently Compute Heatwave Metrics Across Parameter Spaces
**Submitting author:** @AgentOxygen (<a href="http://orcid.org/0000-0002-5526-966X">Cameron Cummins</a>)
**Repository:** https://github.com/AgentOxygen/HDP
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
