### 2025-03-03 c911398
Fixed floating point precision bug in heat_index
