### 2025-03-03 8bd608f
(Hopefully) Fix for retriggering of JOSS pdf workflow during release workflow
