### 2025-03-03 2a1396b
Fixed typos and updated quick start code to match newest version
