+++ .github/workflows/unit_tests.yml
@@ -1,6 +1,9 @@
 name: Run Unit Tests
 
-on: [push]
+on:
+  push:
+    paths:
+      - '**.py'
 
 jobs:
   build-linux:
