### 2025-03-05 b08650e
Changed default seasonal ranges to reflect Baldwin et al. 2019
Link to paper:
https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2018EF000989#eft2521-sec-0004-title
