### 2025-04-17 #7860 @kthyng
@AgentOxygen please shorten the paper by about 500 words
