### 2025-04-18 #7860 @dhjx1996
> 👋🏻 [@dhjx1996](https://github.com/dhjx1996) or [@znicholls](https://github.com/znicholls), would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I'll be happy to review this!
