### 2025-04-18 #7860 @rwegener2
👋🏻 @dvalters, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
