### 2025-04-30 e403bc1
Fixed return type typo
