### 2025-05-23 AgentOxygen/HDP#16 @Chilipp
sample data for usage guide missing
hey! coming from the JOSS review in https://github.com/openjournals/joss-reviews/issues/8111, I wonder whether you have the data for the example usage available. in your example you define some input dir via

```python
input_dir = "/local1/climate_model_output/"
```

but it is not clear, where to get this data and how it should look like. This is relevant for the following two points of the checklist:

- Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
-  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
