### 2025-05-23 #8111 @Chilipp
concerning the following item on the checklist:

> Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

@rwegener2 the paper contains an example figure, is it necessary that the figure can be reproduced to fullfil the JOSS criteria?
