### 2025-05-23 #8111 @Chilipp
@rwegener2, concerning the contribution guidelines. The authors provide the following sentence in the readme

> # Contributing
>
>Please report any bugs, ask questions, and make suggestions through the [GitHub Issues form of this repository](https://github.com/AgentOxygen/HDP/issues).

is this enough to fullfil the criteria?

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

it's probably the shorted contribution guidelines I ever saw, but from my perspective it's enough 😉
