### 2025-05-23 AgentOxygen/HDP#17 @Chilipp
minor formatting issue in the JOSS paper
> just a minor comment from the JOSS review in https://github.com/openjournals/joss-reviews/issues/8111: You should add a newline character here, otherwise the paper does not render well
>
>
>
> ![Image](https://github.com/user-attachments/assets/bbb1f77b-cfaa-4c4b-beee-f1313c2efb4b)

 _Originally posted by @Chilipp in [d95b543](https://github.com/AgentOxygen/HDP/commit/d95b5431be5cd2fab80271a687f8c59cf10671b0#r157841672)_
