### 2025-05-25 #8111 @dhjx1996
Hi @AgentOxygen, I have done a first pass through your software repository and the accompanying paper. Here are my thoughts and some suggestions.

**Statement of Need**

As I understand, and please clarify otherwise, `HDP` is meant to fulfil three needs: 1) Be a convenience tool 2) Be a comprehensive tool 3) Handle "terabyte-scale" data. I am convinced about `HDP`'s value as a convenience tool. I am, however, less convinced on points 2) and 3).

For 2), I would like to see more engagement with the heatwave literature: show that `HDP` can reproduce heatwave plots from NOAA, the IPCC, or research papers. Implement specialized heatwave metrics proposed in some papers. A related point is that the functions can be better explained, for example, how is heat index computed in `hdp.measure.apply_heat_index`? No doubt there are standard definitions like the one used by NOAA, but if so, add the citation.

For 3), a good showcase of `HDP`'s abilities would be to generate a high resolution global map of heatwave frequencies (or some other statistic), something like the global maps in https://doi.org/10.1038/s41467-022-35571-0 perhaps. Could you also add a bit more elaboration on how `HDP` satisfies needs not satisfied by `xclim` and other packages -- `xclim` also uses `xarray` and should in theory be able to manage large datasets. Do properly cite `heatwave3`, `xclim` and `ehfheatwaves`.

**Documentation**

The examples should be fully reproducible with nothing more than cloning the repository and installing `HDP`. Right now the required datasets are missing, thank you @Chilipp for raising the issue. There also needs to be more examples -- this is a good chance to showcase more of `HDP`'s capabilities. In addition, I suggest packaging your examples as Jupyter Notebooks instead of code snippets (why not have both?). Regarding the current _Example 1: Generating Heatwave Diagnostics_, I would suggest showing some plots of the resultant dataset.

As aforementioned, the documentation does not say enough about how the functions are implemented. Preferably, the documentation should give a basic introduction to the study of heatwaves, the usual metrics used and the methods used to compute them, e.g. https://www.epa.gov/climate-indicators/climate-change-indicators-heat-waves.

I will be happy to provide further feedback on software functionality once the examples have been fixed!
