### 2025-05-27 #8111 @rwegener2
Thanks for the clarifying questions @Chilipp!

> the paper contains an example figure, is it necessary that the figure can be reproduced to fullfil the JOSS criteria?

That figure appears to be showing example output from the package. "Original results" generally refers more to analysis done about the package (ex. plotting the outputs of timing runs). In the case of a figure showing package outputs (like we have here) the reproducibility of that figure would be covered more under the "**Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)." bullet. We should be able to reproduce that figure or a similar one if it is the main output of the package, but more due to the requirement for example usage than the requirement for reproducible paper results. Does that make sense?

> it's probably the shorted contribution guidelines I ever saw, but from my perspective it's enough 😉

I would agree that is pretty short. Of the three requirements in the **community guidelines** bullet I think that the text you copied covers "2. Report issues or problems with the software" and "3. Seek support". "1. Contribute to the software" is not covered here. The package should include more of a description of how to produce the technical setup to contribute to the package. This usually involves setting up an environment with an editable installation (ex. `pip install -e .`), installing developer only packages like pytest, and steps for running the tests. The readme is a good place to put an overview, but it's also really common to have "Contributor's Guide" as a page in the documentation. Contributor guidelines can get pretty elaborate (ex. [scipy](https://docs.scipy.org/doc/scipy/dev/contributor/contributor_toc.html)), but in this case I suggest including in a page of the documentation (or the readme) at least 1) a description for how to setup the developer environment and 2) steps to run the tests.

@Chilipp or @AgentOxygen let me know if any of that doesn't make sense.
