### 2025-09-29 #8111 @AgentOxygen
@rwegener2 Apologies for the delay! I expect to have these changes implemented and push a new release by mid-October. I will be sure to notify you when that happens
