### 2025-10-20 e705de6
Added dockerfile for testing
