### 2025-10-20 429448e
Updated time generation to use newer xarray.date_range
