### 2025-10-20 3b2e97e
Fixed bug with dimension ordering
