### 2025-10-20 e7d7b30
Added temporary output directory for testing
