### 2025-10-20 9915be0
Extended default time period to get a better sample
