### 2025-10-31 #8111 @AgentOxygen
@rwegener2 @dhjx1996 Apologies for the confusion, by "issues" I was referring to the other entries in the [Issues](https://github.com/AgentOxygen/HDP/issues) section of the repository (I have added some future enhancements and received feedback from other users outside of JOSS). I believe I have addressed all of the issues raised by the reviewers for JOSS, so please feel free to revisit the package when you find the time! I know this is a busy period for most, so no rush.
