### 2025-11-27 #8111 @dhjx1996
Hi @AgentOxygen, I have looked through your package and paper again and it looks massively improved, congratulations! I also successfully ran your Quick Start code and the results look good. For me these are the remaining issues:

# Paper
- Please cite `heatwave3`, `xclim`, `ehfheatwaves` rather than just link to them.
- I must admit I am still not quite convinced by your claim that "Although software does exist for calculating heatwave metrics (e.g. heatwave3, xclim, ehfheatwaves), these tools are not optimized to analyze more than a few definitions and thresholds at a time nor do they offer diagnostic plots", especially since you do not provide comparison tests or evidence of superior performance. Moreover, I looked at `heatwaveR` and they do have built-in plotting functionality, though perhaps not as automated as yours. I do not think `HDP` must outcompete those packages to be accepted, but I think you should weaken your claims. @rwegener2 @Chilipp any thoughts?

# Package
## Quick Start example
- Let users know what computational resources you used to run each example. I initially tried to run your Quick Start code on my laptop and that did not go well. **I advise including a small example that can be easily run on local machines**, but this is not a requirement.
- The paths in `python hdp/docs/sample_data/sample.py hdp/docs/sample_data/` are slightly wrong; `docs` is outside `hdp`

## LENS2 Heatwave Diagnostics
- I love the idea of this example, but it is unclear to me how to carry it out, in particular, **which exact LENS2 data files do I need**?
- Moreover, how do I know that I carried out the example successfully? It would be good to **plot some of the outputs** so that I can compare results -- and such plots would very much catch the attention of heatwave experts.

## Others
- I ran your pytests successfully and checked off "Automated tests". Three comments though: 1) I would like to have some idea about what is being tested. 2) You should tell users, and not just developers, about the tests. 3) Preferably make the tests automated, using GitHub actions for example.
- The checklist for "Installation instructions" asks "Is there a clearly-stated list of dependencies". I have not seen that. Could you include that somewhere or point me to it?
