### 2026-01-05 #8111 @AgentOxygen
Thank you @dhjx1996! I went through and made modifications to address each of your points. Let me know what you think:

>     * Please cite `heatwave3`, `xclim`, `ehfheatwaves` rather than just link to them.
Done!

>     * I must admit I am still not quite convinced by your claim that "Although software does exist for calculating heatwave metrics (e.g. heatwave3, xclim, ehfheatwaves), these tools are not optimized to analyze more than a few definitions and thresholds at a time nor do they offer diagnostic plots", especially since you do not provide comparison tests or evidence of superior performance. Moreover, I looked at `heatwaveR` and they do have built-in plotting functionality, though perhaps not as automated as yours. I do not think `HDP` must outcompete those packages to be accepted, but I think you should weaken your claims. [@rwegener2](https://github.com/rwegener2) [@Chilipp](https://github.com/Chilipp) any thoughts?
Agreed. I have re-framed that section to present the HDP in _addition_ to these packages as opposed to _over_ these packages. Let me know if the new wording works better.

>     * Let users know what computational resources you used to run each example. I initially tried to run your Quick Start code on my laptop and that did not go well. **I advise including a small example that can be easily run on local machines**, but this is not a requirement.
I decided that the HPC example is likely to niche due to computational constraints. I found a cloud-based database and scaled the example to work with a typical PC. I [updated the documentation](https://hdp.readthedocs.io/en/latest/examples.html#example-analyzing-cesm2-dataset-hosted-in-the-cloud) and added a [Python script](https://github.com/AgentOxygen/HDP/blob/main/docs/example_cmip_workflow/run_cmip_workflow.py) that should work alongside the Docker container. I also dropped the resolution of the quick start example to be more lightweight.

>     * The paths in `python hdp/docs/sample_data/sample.py hdp/docs/sample_data/` are slightly wrong; `docs` is outside `hdp`
Fixed!

>     * I love the idea of this example, but it is unclear to me how to carry it out, in particular, **which exact LENS2 data files do I need**?
Solved using the cloud data described earlier.

>     * Moreover, how do I know that I carried out the example successfully? It would be good to **plot some of the outputs** so that I can compare results -- and such plots would very much catch the attention of heatwave experts.
I added a figure from the figure deck notebook generated by the example. I plan to add an interactive notebook in the documentation itself, but I am still learning how to do that effectively. This is definitely something I can continue to work on.

>     * I ran your pytests successfully and checked off "Automated tests". Three comments though: 1) I would like to have some idea about what is being tested. 2) You should tell users, and not just developers, about the tests. 3) Preferably make the tests automated, using GitHub actions for example.
I added a [section dedicated to unit testing](https://hdp.readthedocs.io/en/latest/testing.html) that describes what each PyTest script is doing, is this enough to inform users in addition to developers? The tests are already automated using GitHub Actions.

>     * The checklist for "Installation instructions" asks "Is there a clearly-stated list of dependencies". I have not seen that. Could you include that somewhere or point me to it?
Done! You can view the list in the documentation [here](https://hdp.readthedocs.io/en/latest/install.html).
