### 2026-01-06 #8111 @dhjx1996
Hi @AgentOxygen, thank you for the changes! I have ticked off the rest of my checklist and I approve the publication of this package and the corresponding paper (@rwegener2).

P.S. @AgentOxygen There, however, seems to be one coding issue in your examples: on my Windows laptop your Quick Start example results in the following error (the example runs fine on my Linux HPC):

```
xception ignored in: <function Image.__del__ at 0x0000020BF9DBED40>
Traceback (most recent call last):
  File "C:\Users\dionh\miniconda3\envs\HDP\Lib\tkinter\__init__.py", line 4258, in __del__
    self.tk.call('image', 'delete', self.name)
RuntimeError: main thread is not in main loop
Exception ignored in: <function Variable.__del__ at 0x0000020BF9D489A0>
Traceback (most recent call last):
  File "C:\Users\dionh\miniconda3\envs\HDP\Lib\tkinter\__init__.py", line 414, in __del__
    if self._tk.getboolean(self._tk.call("info", "exists", self._name)):
RuntimeError: main thread is not in main loop
Exception ignored in: <function Variable.__del__ at 0x0000020BF9D489A0>
Traceback (most recent call last):
  File "C:\Users\dionh\miniconda3\envs\HDP\Lib\tkinter\__init__.py", line 414, in __del__
    if self._tk.getboolean(self._tk.call("info", "exists", self._name)):
RuntimeError: main thread is not in main loop
Exception ignored in: <function Variable.__del__ at 0x0000020BF9D489A0>
Traceback (most recent call last):
  File "C:\Users\dionh\miniconda3\envs\HDP\Lib\tkinter\__init__.py", line 414, in __del__
    if self._tk.getboolean(self._tk.call("info", "exists", self._name)):
RuntimeError: main thread is not in main loop
Exception ignored in: <function Variable.__del__ at 0x0000020BF9D489A0>
Traceback (most recent call last):
  File "C:\Users\dionh\miniconda3\envs\HDP\Lib\tkinter\__init__.py", line 414, in __del__
    if self._tk.getboolean(self._tk.call("info", "exists", self._name)):
RuntimeError: main thread is not in main loop
Tcl_AsyncDelete: async handler deleted by the wrong thread
```

I got **Claude** to explain the error and the solution, which is to switch from the interactive `tkinter` backend to the `Agg` backend:

**The Fundamental Issue**

* Interactive backends (like TkAgg) create GUI resources - tkinter windows, event loops, etc.
* GUI toolkits require their resources to be managed by the main thread - this is a fundamental constraint of most GUI frameworks (tkinter, Qt, etc.)
* Your script creates matplotlib figures but exits immediately - without ever running an interactive event loop
* During Python shutdown, matplotlib tries to clean up these GUI resources - but the cleanup happens in the wrong thread context or after the GUI event loop has already been torn down
* Result: "RuntimeError: main thread is not in main loop" - tkinter objects are being destroyed, but the main event loop that manages them isn't running (or is already gone)

**Why Agg solves it**

The `Agg` backend:

* Creates no GUI resources at all
* Renders directly to memory/files
* Has nothing that requires thread-specific cleanup
* No event loop needed

**Analogy**
* It's like trying to close a window in a house that's already been demolished - the window cleanup code expects the house (event loop) to still exist, but it's gone.
* Your understanding is spot-on. This is a classic issue with GUI backends in non-interactive scripts, and switching to a non-GUI backend is the standard solution.

I made a small pull request to switch both example scripts to the `Agg` backend and with that change both examples work on both my Linux HPC and Windows laptop.
