### 2026-01-21 #8111 @Chilipp
@rwegener2 I am done with my review and everything worked out well. Just a minor fix in https://github.com/AgentOxygen/HDP/pull/26 to correct for the package name. I recommend this paper for publication. My apologies again @AgentOxygen for the delay of the review, the days before christmas where just too much packed with other stuff
