### 2026-01-22 #8111 @rwegener2
Thanks @Chilipp and @dhjx1996 for your thoughtful reviews on this submission! JOSS would not be possible without you.

@AgentOxygen at this point I’m going to move into my final set of checks. I will go through a few steps and then ask you to follow up on some steps for creating and archive and DOI.
