### 2026-01-22 #8111 @rwegener2
Hey @AgentOxygen, thanks for writing a nice paper! I just have a few comments that I think could help make the Statement of Need clearer:
- line 27: “which have remained mostly specific to individual studies” The meaning is not clear to me. Are you saying that only a few studies have chosen to use data from large ensembles?
- line 28: “this problem” — You mention several problems throughout the first paragraph, and I get a bit lost by the end which problem you are addressing. Some thoughts that could help:
1. perhaps some additional detail on “generalized tools”
2. A list (either bulleted or inline) could perhaps help? It seems like this package is solving several problems: (1) “there are no centralized tools or universal metrics” (2) “excessive data management and computational burden of sampling a greater number of parameters” (3) “large ensembles”.
3. Alternately, removing “however” in line 24 or using a transition like “While studies could attempt to report metrics across parameter space, excessive data management and computational burden … make it difficult for studies to report more than only a handful of metrics and definitions"
One note that may help you understand my perspective as a reader is that I thought the first sentence of your second paragraph was really clear, and I think your first paragraph made more sense to me after reading the first sentence of the second paragraph.

That feedback is pretty open ended, so please don’t hesitate to ask if you have any questions! Once you have a chance to address that you can work through the author action items in step 2 of the [checklist I posted above](https://github.com/openjournals/joss-reviews/issues/8111#issuecomment-3785317648).
