### 2026-01-22 #8111 @AgentOxygen
@rwegener2 Thanks for the feedback! I made some changes that I think improve it a bit. Let me know what you think:

https://github.com/AgentOxygen/HDP/commit/6511cca50a0423515ee88ca378e7e9f8134f01a8
