### 2026-01-23 #8111 @AgentOxygen
@rwegener2

The Zenodo archive: https://doi.org/10.5281/zenodo.18354742

The package release is up to date, all feedback was related to documentation and example codes: https://github.com/AgentOxygen/HDP/releases/tag/v1.0.2

I also added the ORCIDs for the authors. Let me know if there is anything else I need to do!
