### 2026-01-23 #8111 @rwegener2
Thanks @AgentOxygen! I noticed that the title of your zenodo archive and the title of your paper are different. We typically recommend that those titles are the same, since they make one unified collection. If you have a motivation for making them different, however, we can be flexible.

Could you please either change the title of the Zenodo archive to match the paper or let me know your reason for making them different?
