### 2026-01-24 #8111 @rwegener2
Alright @AgentOxygen, I’ve recommended your paper for acceptance! At this point the track editor will review the submission and complete the final steps. Thanks for submitting to JOSS!
