### 2026-01-24 #8111 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
